+++ tests/Stubs/DriverStub.php
@@ -1,0 +1,7 @@
+<?php
+namespace Michaels\Spider\Test\Stubs;
+
+class DriverStub
+{
+
+}