+++ CHANGELOG.md
@@ -4,5 +4,7 @@ All Notable changes to `Spider` will be documented in this file
 
 ## v0.1 - NEXT
 ### Added
+- Initialized contracts: DriverInterface and ConnectionInterface
+- Setup connection manager
 - Init Repo
 - Initial Tests